+++ src/dolphin/gx/GXPixel.h
@@ -15,12 +15,12 @@ void GXInitFogAdjTable(GXFogAdjTbl*, u16, Mtx44);
 void GXSetFogRangeAdj(s8, u16, GXFogAdjTbl*);
 void GXSetPixelFmt(int, int);
 void GXSetFieldMask(s32, s8);
-void GXSetDstAlpha(s32, s32);
-void GXSetDither(s32);
-void GXSetZCompLoc(s32);
-void GXSetZMode(s8, s32, s32);
-void GXSetAlphaUpdate(s32);
-void GXSetColorUpdate(s32);
+void GXSetDstAlpha(u8, u8);
+void GXSetDither(u8);
+void GXSetZCompLoc(u8);
+void GXSetZMode(u8, u32, u32);
+void GXSetAlphaUpdate(u8);
+void GXSetColorUpdate(u8);
 void GXSetBlendMode(s32, s32, s32, s32);
 void GXSetFog(s32, GXColor, f32, f32, f32, f32);
